requirement.txt
apturl==0.5.2
asgiref==3.5.2
asttokens==2.0.5
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.11.1
blinker==1.4
Brlapi==0.7.0
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==2.0.12
cliapp==1.20180812.1
Click==7.0
cmdtest==0.32+git
colorama==0.4.3
command-not-found==0.3
commonmark==0.9.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.0.4
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-crontab==0.7.1
# Editable Git install with no remote (django-debug-toolbar==3.3.0)
-e /home/intern3/Desktop/python/poll_app/src/django-debug-toolbar
django-extensions==3.1.5
django-filter==21.1
django-rest-auth==0.9.5
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
drf-yasg==1.20.0
duplicity==0.8.12.0
ecdsa==0.17.0
entrypoints==0.3
executing==0.8.3
fasteners==0.14.1
future==0.18.2
greenlet==0.4.15
html5lib==1.1
httpie==3.2.1
httplib2==0.14.0
idna==2.8
importlib-metadata==4.11.3
inflection==0.5.1
ipdb==0.13.9
ipython==8.3.0
ipython_genutils==0.2.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
monotonic==1.5
msgpack==0.6.2
multidict==6.0.2
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
packaging==20.3
paramiko==2.6.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
prompt-toolkit==3.0.29
protobuf==3.6.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pycairo==1.16.2
pycups==1.9.73
Pygments==2.12.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pynvim==0.4.1
pyparsing==2.4.6
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-jose==3.3.0
pytz==2022.1
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
rich==12.4.2
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
SecretStorage==2.3.1
simplejson==3.16.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.2.0
systemd-python==234
toml==0.10.2
traitlets==5.2.1.post0
ttystatus==0.38
typing_extensions==4.2.0
ubuntu-advantage-tools==27.8
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.25.8
usb-creator==0.3.7
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
xkit==0.0.0
zipp==3.8.0
